+++ test/integ/allocation-timeout/integ.allocation-timeout.ts.snapshot/atmosphere-integ-allocation-timeout-assertions.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "39.0.0",
   "files": {
-    "0524071f7a081ff0ef7c96d872e55dcef3ca18825fca8956441a16034e37404c": {
+    "d82605528b068001650d0e8693250ccf404b629e8f0e34d692e2f4eb3b84f4c6": {
       "source": {
-        "path": "asset.0524071f7a081ff0ef7c96d872e55dcef3ca18825fca8956441a16034e37404c",
+        "path": "asset.d82605528b068001650d0e8693250ccf404b629e8f0e34d692e2f4eb3b84f4c6",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "0524071f7a081ff0ef7c96d872e55dcef3ca18825fca8956441a16034e37404c.zip",
+          "objectKey": "d82605528b068001650d0e8693250ccf404b629e8f0e34d692e2f4eb3b84f4c6.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -27,7 +27,7 @@
         }
       }
     },
-    "f8a5e662ddfac34bc49bbccb1e6cd6e037318576baa4666358b5eb42b26f2682": {
+    "4aacffa83cccc2b57b6c15dedd8b2d0e27c16248907a980639128a79166629f2": {
       "source": {
         "path": "atmosphere-integ-allocation-timeout-assertions.template.json",
         "packaging": "file"
@@ -35,7 +35,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "f8a5e662ddfac34bc49bbccb1e6cd6e037318576baa4666358b5eb42b26f2682.json",
+          "objectKey": "4aacffa83cccc2b57b6c15dedd8b2d0e27c16248907a980639128a79166629f2.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
