+++ test/integ/allocation-timeout/integ.allocation-timeout.ts.snapshot/tree.json
@@ -1296,9 +1296,9 @@
                       "version": "2.177.0"
                     }
                   },
-                  "AdminRole${Token[AWS.AccountId.0]}eu-central-1": {
-                    "id": "AdminRole${Token[AWS.AccountId.0]}eu-central-1",
-                    "path": "atmosphere-integ-allocation-timeout/Atmosphere/Cleanup/AdminRole${Token[AWS.AccountId.0]}eu-central-1",
+                  "AdminRole${Token[AWS.AccountId.1]}eu-central-1": {
+                    "id": "AdminRole${Token[AWS.AccountId.1]}eu-central-1",
+                    "path": "atmosphere-integ-allocation-timeout/Atmosphere/Cleanup/AdminRole${Token[AWS.AccountId.1]}eu-central-1",
                     "constructInfo": {
                       "fqn": "aws-cdk-lib.Resource",
                       "version": "2.177.0"
@@ -2239,9 +2239,9 @@
                       "version": "2.177.0"
                     }
                   },
-                  "AdminRole${Token[AWS.AccountId.0]}eu-central-1": {
-                    "id": "AdminRole${Token[AWS.AccountId.0]}eu-central-1",
-                    "path": "atmosphere-integ-allocation-timeout/Atmosphere/Allocate/AdminRole${Token[AWS.AccountId.0]}eu-central-1",
+                  "AdminRole${Token[AWS.AccountId.1]}eu-central-1": {
+                    "id": "AdminRole${Token[AWS.AccountId.1]}eu-central-1",
+                    "path": "atmosphere-integ-allocation-timeout/Atmosphere/Allocate/AdminRole${Token[AWS.AccountId.1]}eu-central-1",
                     "constructInfo": {
                       "fqn": "aws-cdk-lib.Resource",
                       "version": "2.177.0"
@@ -3652,7 +3652,7 @@
                       "s3Bucket": {
                         "Fn::Sub": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}"
                       },
-                      "s3Key": "0524071f7a081ff0ef7c96d872e55dcef3ca18825fca8956441a16034e37404c.zip"
+                      "s3Key": "d82605528b068001650d0e8693250ccf404b629e8f0e34d692e2f4eb3b84f4c6.zip"
                     },
                     "description": "test/integ/allocation-timeout/assert.lambda.ts",
                     "environment": {
