+++ test/integ/cleanup-timeout/integ.cleanup-timeout.ts.snapshot/atmosphere-integ-cleanup-timeout-assertions.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "39.0.0",
   "files": {
-    "180849294c473acfbae0524276b002208dd4a6dccf01ccebb7fb00452c31b995": {
+    "54c527bcba4638ae4539c13bd5d5df530440318d89f83d94847ec0a0a973c8c1": {
       "source": {
-        "path": "asset.180849294c473acfbae0524276b002208dd4a6dccf01ccebb7fb00452c31b995",
+        "path": "asset.54c527bcba4638ae4539c13bd5d5df530440318d89f83d94847ec0a0a973c8c1",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "180849294c473acfbae0524276b002208dd4a6dccf01ccebb7fb00452c31b995.zip",
+          "objectKey": "54c527bcba4638ae4539c13bd5d5df530440318d89f83d94847ec0a0a973c8c1.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -40,7 +40,7 @@
         }
       }
     },
-    "421494ca111914732228e1cc2d43b485eb2b24cd96aec76c78ec17522ac4dd14": {
+    "12c15815eb71b122a72ef87e9b1a40571791974f774321f628681d39067c797f": {
       "source": {
         "path": "atmosphere-integ-cleanup-timeout-assertions.template.json",
         "packaging": "file"
@@ -48,7 +48,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "421494ca111914732228e1cc2d43b485eb2b24cd96aec76c78ec17522ac4dd14.json",
+          "objectKey": "12c15815eb71b122a72ef87e9b1a40571791974f774321f628681d39067c797f.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
