+++ test/integ/cleanup-timeout/integ.cleanup-timeout.ts.snapshot/atmosphere-integ-cleanup-timeout-assertions.template.json
@@ -107,7 +107,7 @@
      "S3Bucket": {
       "Fn::Sub": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}"
      },
-     "S3Key": "180849294c473acfbae0524276b002208dd4a6dccf01ccebb7fb00452c31b995.zip"
+     "S3Key": "54c527bcba4638ae4539c13bd5d5df530440318d89f83d94847ec0a0a973c8c1.zip"
     },
     "Description": "test/integ/cleanup-timeout/assert.lambda.ts",
     "Environment": {
@@ -212,7 +212,7 @@
      }
     },
     "flattenResponse": "false",
-    "salt": "1738508565781"
+    "salt": "1738509482963"
    },
    "UpdateReplacePolicy": "Delete",
    "DeletionPolicy": "Delete"
