+++ test/integ/cleanup/integ.cleanup.ts.snapshot/atmosphere-integ-cleanup-assertions.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "39.0.0",
   "files": {
-    "ac05b2cda1c88c69fe009be40514c6805458069ac0e89da21e2964f4f986b321": {
+    "9d537ef66c7487357360da40fe6a9edff3cbef53778af1cd7abc1d64ba2e5c56": {
       "source": {
-        "path": "asset.ac05b2cda1c88c69fe009be40514c6805458069ac0e89da21e2964f4f986b321",
+        "path": "asset.9d537ef66c7487357360da40fe6a9edff3cbef53778af1cd7abc1d64ba2e5c56",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "ac05b2cda1c88c69fe009be40514c6805458069ac0e89da21e2964f4f986b321.zip",
+          "objectKey": "9d537ef66c7487357360da40fe6a9edff3cbef53778af1cd7abc1d64ba2e5c56.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -40,7 +40,7 @@
         }
       }
     },
-    "3fbe1337033dcf7f33caf394540cad51ec6c7296eecbc5071594d9a76c3033d3": {
+    "c984e543a5757ff154cab0dc623bf13b7e7f98ccb057feb1c02cb3de1878723d": {
       "source": {
         "path": "atmosphere-integ-cleanup-assertions.template.json",
         "packaging": "file"
@@ -48,7 +48,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "3fbe1337033dcf7f33caf394540cad51ec6c7296eecbc5071594d9a76c3033d3.json",
+          "objectKey": "c984e543a5757ff154cab0dc623bf13b7e7f98ccb057feb1c02cb3de1878723d.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
