+++ test/integ/deallocate/integ.deallocate.ts.snapshot/atmosphere-integ-deallocate-assertions.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "39.0.0",
   "files": {
-    "6797f9ff67f5d172fe146a454a61d435a1e4026b6ddf71de941f45936e7a606a": {
+    "2cfcca0af76e032c185988b913f39798ca505019488386eb1d5a59f5306fdba8": {
       "source": {
-        "path": "asset.6797f9ff67f5d172fe146a454a61d435a1e4026b6ddf71de941f45936e7a606a",
+        "path": "asset.2cfcca0af76e032c185988b913f39798ca505019488386eb1d5a59f5306fdba8",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "6797f9ff67f5d172fe146a454a61d435a1e4026b6ddf71de941f45936e7a606a.zip",
+          "objectKey": "2cfcca0af76e032c185988b913f39798ca505019488386eb1d5a59f5306fdba8.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -27,7 +27,7 @@
         }
       }
     },
-    "2b1e9292abd6d474db1ba2e57168e8dfa336e22fb1a5b4118f2f5f66de4ad416": {
+    "56964222873a325bafb58a89cf6bb872d9dcdcbc18097dc1165f84a1c3f3eeb5": {
       "source": {
         "path": "atmosphere-integ-deallocate-assertions.template.json",
         "packaging": "file"
@@ -35,7 +35,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "2b1e9292abd6d474db1ba2e57168e8dfa336e22fb1a5b4118f2f5f66de4ad416.json",
+          "objectKey": "56964222873a325bafb58a89cf6bb872d9dcdcbc18097dc1165f84a1c3f3eeb5.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
