+++ test/integ/dev/integ.dev.ts.snapshot/atmosphere-integ-dev-assertions.assets.json
@@ -1,15 +1,15 @@
 {
   "version": "39.0.0",
   "files": {
-    "39bba9e59e07df6c8db55fbeaa33ce462ef67fd1e406413565af49e1b41cee2a": {
+    "39d1e1395a61db34697924460bd81d9009f8e37a44cda1a3060e3993224bca8c": {
       "source": {
-        "path": "asset.39bba9e59e07df6c8db55fbeaa33ce462ef67fd1e406413565af49e1b41cee2a",
+        "path": "asset.39d1e1395a61db34697924460bd81d9009f8e37a44cda1a3060e3993224bca8c",
         "packaging": "zip"
       },
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "39bba9e59e07df6c8db55fbeaa33ce462ef67fd1e406413565af49e1b41cee2a.zip",
+          "objectKey": "39d1e1395a61db34697924460bd81d9009f8e37a44cda1a3060e3993224bca8c.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -27,7 +27,7 @@
         }
       }
     },
-    "42172f9e88a793c36aea4af39e693695ce6ee52fcdcb09729b9f2e8dd30ce6a3": {
+    "22c28485e71dadae991791d0455770d3c1b9ba50c8f2b2929b40721d591a162d": {
       "source": {
         "path": "atmosphere-integ-dev-assertions.template.json",
         "packaging": "file"
@@ -35,7 +35,7 @@
       "destinations": {
         "current_account-current_region": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "42172f9e88a793c36aea4af39e693695ce6ee52fcdcb09729b9f2e8dd30ce6a3.json",
+          "objectKey": "22c28485e71dadae991791d0455770d3c1b9ba50c8f2b2929b40721d591a162d.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
